First draft of a quick cheat sheet. Please feel free to send feedback in the form of comments, suggestions, error catches, etc. or just edit the document yourself.
This documents brieﬂy explains the main guidelines followed by the
pro jects PySAL and GeoDaSpace to handle their in-code documentation
and testing modules. The main structures and mark-up are outlined as
well as the principal tricks to know in order to write doctests that may
be added to the global testing system.

To compile in LaTeX, run the following command twice:

> pdflatex -shell-escape howto_contrib2cookbook.tex